refactor: Moved the moderator screen to account settings and the MFA switch to preferences. I also made the moderator screen refresh upon mounting to avoid having to reload the app.
I moved the moderator screen to reduce clutter in the planets/tabstack, and the settings seemed like a more logical place for role specific actions. I moved the MFA switch as it was awkward amongst the other settings and Dark Mode was the only thing under preferences.
